Update timed tests & testing threaded tests
