+++ .gitignore
@@ -14,3 +14,5 @@ dist/
 build/
 
 .pytest_cache/
+
+tests/__pycache__
